+++ processor/src/test/fixtures/bad_input/com/example/render/UnexpectedRenderComponent.java
@@ -1,0 +1,15 @@
+package com.example.render;
+
+import react4j.ReactNode;
+import react4j.annotations.Render;
+import react4j.annotations.View;
+
+@View( requireRender = false )
+abstract class UnexpectedRenderComponent
+{
+  @Render
+  ReactNode render()
+  {
+    return null;
+  }
+}