+++ processor/src/test/fixtures/input/com/example/render/NoRenderWithPostConstructView.java
@@ -1,0 +1,13 @@
+package com.example.render;
+
+import arez.annotations.PostConstruct;
+import react4j.annotations.View;
+
+@View( requireRender = false )
+abstract class NoRenderWithPostConstructView
+{
+  @PostConstruct
+  void postConstruct()
+  {
+  }
+}