+++ processor/src/test/fixtures/input/com/example/render/NoRenderWithPostMountOrUpdateView.java
@@ -1,0 +1,13 @@
+package com.example.render;
+
+import react4j.annotations.PostMountOrUpdate;
+import react4j.annotations.View;
+
+@View( requireRender = false )
+abstract class NoRenderWithPostMountOrUpdateView
+{
+  @PostMountOrUpdate
+  void postMountOrUpdate()
+  {
+  }
+}