+++ processor/src/test/fixtures/input/com/example/render/NoRenderWithPostMountView.java
@@ -1,0 +1,13 @@
+package com.example.render;
+
+import react4j.annotations.PostMount;
+import react4j.annotations.View;
+
+@View( requireRender = false )
+abstract class NoRenderWithPostMountView
+{
+  @PostMount
+  void postMount()
+  {
+  }
+}