+++ processor/src/test/fixtures/input/com/example/render/NoRenderWithPreUpdateView.java
@@ -1,0 +1,13 @@
+package com.example.render;
+
+import react4j.annotations.PreUpdate;
+import react4j.annotations.View;
+
+@View( requireRender = false )
+abstract class NoRenderWithPreUpdateView
+{
+  @PreUpdate
+  void preUpdate()
+  {
+  }
+}