+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Generate an error if a view contains a `@ScheduleRender` annotated method and is not of type `STATEFUL`.
+* Allow `@View` instances to omit th `@Render` annotated method if they specify `@View(requireRender=false)`
 
 ### [v0.168](https://github.com/react4j/react4j/tree/v0.168) (2020-06-25) · [Full Changelog](https://github.com/react4j/react4j/compare/v0.167...v0.168)